Move `cleanJSON` to `./src/utils`
Now with this function unified and reusuable at the two callsites, we
can move this to the utils module and import it instead.

Note, that i've included the function as named in the inline definition.
This is to simplify the recursive call.
