+++ ImageMarkupBuilder.js
@@ -1,3 +1,5 @@
+const { cleanJSON } = require('./src/utils');
+
 var isNode = typeof window == 'undefined';
 
 /**
@@ -78,42 +80,6 @@ function ImageMarkupBuilder(fabricCanvas) {
               rect.left + (strokeWidth * 2) > this.width ||
               rect.top + rect.height - (strokeWidth * 2) < 0 ||
               rect.top + (strokeWidth * 2) > this.height);
-   }
-
-   /**
-    * Cycles through an object and changes all numeric fields to ints
-    * where necessary. 'context' is used for exception reporting and can be
-    * left unset upon invocation.
-    */
-   function cleanJSON(json, context) {
-      if (!context)
-         context = "root";
-
-      var integerProperties = [
-         'x'
-        ,'y'
-        ,'width'
-        ,'height'
-        ,'radius'
-        ,'strokeWidth'
-      ];
-      const isIntegerProperty = (property) => integerProperties.indexOf(property) != -1;
-
-      for (var property in json) {
-         if (typeof(json[property]) == 'object') {
-            cleanJSON(json[property], context + '.' + property);
-         }
-         else if (isIntegerProperty(property)) {
-            if (typeof(json[property]) == 'string') {
-             json[property] = parseInt(json[property]);
-               if (isNaN(json[property])) {
-                  var msg = "In '" + context + "': property '" + property +
-                   "' is not a number.";
-                  throw msg;
-               }
-            }
-         }
-      }
    }
 
    function applyBackground(callback) {
