+++ ImageMarkupCall.js
@@ -1,4 +1,4 @@
-var { Int } = require('./src/utils');
+var { Int, cleanJSON } = require('./src/utils');
 var ImageMarkupBuilder = require('./ImageMarkupBuilder').Builder;
 var Fabric = require('fabric').fabric;
 var argv = require('optimist').argv;
@@ -213,42 +213,6 @@ function processJSON(json) {
          console.log(builder.getMarkupString());
       }
    });
-}
-
-/**
- * Cycles through an object and changes all numeric fields to ints
- * where necessary. 'context' is used for exception reporting and can be
- * left unset upon invocation.
- */
-function cleanJSON(json, context) {
-   if (!context)
-      context = "root";
-
-   var integerProperties = [
-      'x'
-     ,'y'
-     ,'width'
-     ,'height'
-     ,'radius'
-     ,'strokeWidth'
-   ];
-   const isIntegerProperty = (property) => integerProperties.indexOf(property) != -1;
-
-   for (var property in json) {
-
-      if (typeof(json[property]) == 'object') {
-         cleanJSON(json[property], context + '.' + property);
-      } else if (isIntegerProperty(property)) {
-         if (typeof(json[property]) == 'string') {
-            json[property] = Int(json[property]);
-            if (isNaN(json[property])) {
-               var msg = "In '" + context + "': property '" + property
-                + "' is not a number.";
-               throw msg;
-            }
-         }
-      }
-   }
 }
 
 processArgs();
